Docs/SK-901: Fixes related to README format (#646)
