+++ examples/monai-2D-mednist/README.rst
@@ -1,15 +1,15 @@
 FEDn Project: MonAI 2D Classification with the MedNIST Dataset (PyTorch)
 ------------------------------------------------------------------------
 
-This is an example FEDn Project based on the  MonAI 2D Classification with the MedNIST Dataset. 
+This is an example FEDn Project based on the  MonAI 2D Classification with the MedNIST Dataset.
 The example is intented as a minimalistic quickstart and automates the handling of training data
-by letting the client download and create its partition of the dataset as it starts up. 
+by letting the client download and create its partition of the dataset as it starts up.
 
-Links: 
-   
+Links:
+
 -  MonAI: https://monai.io/
 -  Base example notebook: https://github.com/Project-MONAI/tutorials/blob/main/2d_classification/mednist_tutorial.ipynb
--  MedNIST dataset: https://github.com/Project-MONAI/MONAI-extra-test-data/releases/download/0.8.1/MedNIST.tar.gz 
+-  MedNIST dataset: https://github.com/Project-MONAI/MONAI-extra-test-data/releases/download/0.8.1/MedNIST.tar.gz
 
 Prerequisites
 -------------
@@ -17,17 +17,18 @@ Prerequisites
 Using FEDn Studio:
 
 -  `Python 3.8, 3.9, 3.10 or 3.11 <https://www.python.org/downloads>`__
--  `A FEDn Studio account <https://fedn.scaleoutsystems.com/signup>`__   
+-  `A FEDn Studio account <https://fedn.scaleoutsystems.com/signup>`__
 
 If using pseudo-distributed mode with docker-compose:
 
 -  `Docker <https://docs.docker.com/get-docker>`__
 -  `Docker Compose <https://docs.docker.com/compose/install>`__
 
+
 Creating the compute package and seed model
 -------------------------------------------
 
-Install fedn: 
+Install fedn:
 
 .. code-block::
 
@@ -54,13 +55,31 @@ Next, generate a seed model (the first model in a global model trail):
 
    fedn run build --path client
 
-This will create a seed model called 'seed.npz' in the root of the project. This step will take a few minutes, depending on hardware and internet connection (builds a virtualenv).  
+This will create a seed model called 'seed.npz' in the root of the project. This step will take a few minutes, depending on hardware and internet connection (builds a virtualenv).
+
+Download and Prepare the data
+-------------------------------------------
+
+Install requirements:
+
+.. code-block::
+
+   pip install -r requirements.txt
+
+Download and divide the data into parts. Set the number of
+data parts as an arguments python prepare_data.py NR-OF-DATAPARTS. In the
+below command we divide the dataset into 10 parts.
+.. code-block::
+
+    python prepare_data.py 10
+
+
 
 Using FEDn Studio
 -----------------
 
 Follow the guide here to set up your FEDn Studio project and learn how to connect clients (using token authentication): `Studio guide <https://fedn.readthedocs.io/en/stable/studio.html>`__.
-On the step "Upload Files", upload 'package.tgz' and 'seed.npz' created above. 
+On the step "Upload Files", upload 'package.tgz' and 'seed.npz' created above.
 
 Connecting clients:
 ===================
@@ -70,23 +89,27 @@ Connecting clients:
 .. code-block::
 
    export FEDN_PACKAGE_EXTRACT_DIR=package
-   export FEDN_DATA_PATH=./data/
+   export FEDN_DATA_PATH=<full_path_to_the_dir>/data/
    export FEDN_CLIENT_SETTINGS_PATH=<full_path_to_the_dir>/client_settings.yaml
+   export FEDN_DATA_SPLIT_INDEX=0
+
    fedn client start -in client.yaml --secure=True --force-ssl
 
 Connecting clients using Docker:
 ================================
 
-For convenience, there is a Docker image hosted on ghrc.io with fedn preinstalled. To start a client using Docker: 
+For convenience, there is a Docker image hosted on ghrc.io with fedn preinstalled. To start a client using Docker:
 
 .. code-block::
 
    docker run \
      -v $PWD/client.yaml:/app/client.yaml \
      -v $PWD/client_settings.yaml:/app/client_settings.yaml \
+     -v $PWD/data:/app/data \
      -e FEDN_PACKAGE_EXTRACT_DIR=package \
-     -e FEDN_DATA_PATH=./data/ \
+     -e FEDN_DATA_PATH=/app/data/ \
      -e FEDN_CLIENT_SETTINGS_PATH=/app/client_settings.yaml \
+     -e FEDN_DATA_SPLIT_INDEX=0 \
      ghcr.io/scaleoutsystems/fedn/fedn:0.9.0 run client -in client.yaml --force-ssl --secure=True
 
 
@@ -107,8 +130,8 @@ Start a pseudo-distributed FEDn network using docker-compose:
     -f docker-compose.override.yaml \
     up
 
-This starts up local services for MongoDB, Minio, the API Server, one Combiner and two clients. 
-You can verify the deployment using these urls: 
+This starts up local services for MongoDB, Minio, the API Server, one Combiner and two clients.
+You can verify the deployment using these urls:
 
 - API Server: http://localhost:8092/get_controller_status
 - Minio: http://localhost:9000
@@ -123,18 +146,18 @@ Upload the package and seed model to FEDn controller using the APIClient. In Pyt
    client.set_active_package("package.tgz", helper="numpyhelper")
    client.set_active_model("seed.npz")
 
-You can now start a training session with 5 rounds (default): 
+You can now start a training session with 5 rounds (default):
 
 .. code-block::
 
    client.start_session()
 
-Automate experimentation with several clients  
+Automate experimentation with several clients
 =============================================
 
-If you want to scale the number of clients, you can do so by modifying ``docker-compose.override.yaml``. For example, 
-in order to run with 3 clients, change the environment variable ``FEDN_NUM_DATA_SPLITS`` to 3, and add one more client 
-by copying ``client1`` and setting ``FEDN_DATA_PATH`` to ``/app/package/data3/``
+If you want to scale the number of clients, you can do so by modifying ``docker-compose.override.yaml``. For example,
+in order to run with 3 clients, change the environment variable ``FEDN_NUM_DATA_SPLITS`` to 3, and add one more client
+by copying ``client1``.
 
 
 Access message logs and validation data from MongoDB
