+++ examples/monai-2D-mednist/client/python_env.yaml
@@ -2,11 +2,11 @@ name: monai-2d-mdnist
 build_dependencies:
   - pip
   - setuptools
-  - wheel==0.37.1
+  - wheel
 dependencies:
   - torch==2.2.1
   - torchvision==0.17.1
-  - fedn==0.9.0
+  - fedn
   - monai-weekly[pillow, tqdm]
-  - scikit-learn
-  - tensorboard   
+  - numpy==1.26.4
+  - scikit-learn
