+++ examples/monai-2D-mednist/client_settings.yaml
@@ -1,6 +1,5 @@
 lr: 0.01
-batch_size: 32
-local_epochs: 10
+batch_size: 8
+local_epochs: 1
 num_workers: 1
 sample_size: 30
-split_index: 4
