+++ examples/monai-2D-mednist/requirements.txt
@@ -1,0 +1,3 @@
+monai
+PyYAML
+numpy==1.26.4
